chore(AppArg): Move from bin/hrtor.rs to cli/src/lib.rs
