+++ Cargo.toml
@@ -19,6 +19,7 @@ clap = { version = "4.4.7", features = ["derive"] }
 linefeed = "0.6.0"
 rlua = "0.19.7"
 file_loader = { path = "file_loader" }
+cli = { path = "cli" }
 
 [workspace]
