+++ cli/Cargo.toml
@@ -4,3 +4,5 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
+clap = { version = "4.4.7", features = ["derive"] }
+file_loader = { path = "../file_loader" }
