+++ src/bin/hrtor.rs
@@ -1,47 +1,13 @@
 use clap::Parser;
 use file_loader::{CommandLineArgsParser, FileInfo};
 use hrtor::{constants::PROMPT, CommandResult, CommandStatus, Hrtor, HrtorProcessor};
+use cli::AppArg;
 
 use linefeed::Interface;
 use std::{
     error::Error,
     sync::{Arc, Mutex},
 };
-
-#[derive(Parser)]
-struct AppArg {
-    /// File's Path
-    #[arg(help = "The file you want to edit")]
-    pub path: String,
-
-    //#[arg(long, default_value_t = String::from("./init.lua"))]
-    #[arg(short, long, default_value_t = String::from("./init.lua"), help = "your config file which is as config.lua")]
-    pub config: String,
-}
-
-impl CommandLineArgsParser for AppArg {
-    fn read_fileinfo(&self) -> Result<FileInfo, Box<dyn Error>> {
-        Ok(FileInfo {
-            path: self.path.clone(),
-            context: std::fs::read_to_string(self.path.clone()).unwrap_or_else(|_| {
-                eprintln!("your file cannot find. create a new buffer to continue this process.");
-                String::new()
-            }),
-        })
-    }
-
-    fn read_configinfo(&self) -> Result<FileInfo, Box<dyn Error>> {
-        Ok(FileInfo {
-            path: self.config.clone(),
-            context: std::fs::read_to_string(self.config.clone()).unwrap_or_else(|_| {
-                eprintln!(
-                    "your config file cannot find. Continue this process without config file."
-                );
-                String::new()
-            }),
-        })
-    }
-}
 
 /// main function
 fn main() -> Result<(), Box<dyn Error>> {
